Moved reusable CLI classes to their own namespace
