+++ src/MigrateRelationshipsCliCommand.php
@@ -6,6 +6,7 @@ namespace Inpsyde\MultilingualPress2to3;
 use Dhii\I18n\FormatTranslatorInterface;
 use Dhii\I18n\StringTranslatorConsumingTrait;
 use Exception;
+use Inpsyde\MultilingualPress2to3\Cli\CliUtilsTrait;
 use Throwable;
 use wpdb as Wpdb;
